+++ webapp/frontend/components/utils/commands.js
@@ -7,6 +7,7 @@ class Command {
         this.label = props.label;
         this.subcommands = props.subcommands ? props.subcommands : []; 
         this.action = props.action;
+        this.args_stack = [] // forwarded, not consumed args
     }
 
     get_options() {
@@ -20,12 +21,31 @@ class Command {
     }
 }
 
-
 const commands_config = [
     new Command({
         id: 'start',
         label: 'Start',
         subcommands: [
+            new Command({
+                id: 'add_bucket',
+                label: 'Add bucket',
+                prefix: 'Link',
+                subcommands: [
+                    new Command({
+                        id: ON_ENTER_COMMAND,
+                        prefix: 'Name',
+                        subcommands: [
+                            new Command({
+                                id: ON_ENTER_COMMAND,
+                                action: (api, name, args_stack) => { 
+                                    let url = args_stack.pop();
+                                    api.onSave(name, url);
+                                }
+                            })
+                        ]
+                    })
+                ]
+            }),
             new Command({
                 id: 'open_website',
                 label: 'Open website',
@@ -35,12 +55,12 @@ const commands_config = [
             }),
             new Command({
                 id: 'search',
-                prefix: 'search',
                 label: 'Search app',
+                prefix: 'search',
                 subcommands: [
                     new Command({
                         id: ON_ENTER_COMMAND,
-                        action: (text, api) => { api.onSearch(text); }
+                        action: (api, app_name) => { api.onSearch(app_name); }
                     })
                 ]
             })